pyproject.toml
[tool.poetry]
name = "youtube-caption-generator"
version = "0.1.0"
description = ""
authors = ["sucpark <sucpark0317@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numba = "0.61.0"
yt-dlp = "^2025.6.9"
elevenlabs = "^2.3.0"
python-dotenv = "^1.1.0"
openai = "^1.86.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
